refactor(bundle-size)!: extract Tab keydown handler into opt-in module (#549)
Moves Tab-key handling, next-tabbable dispatch, and tabbing state into an
opt-in module installed by getRootDummyInputs.

Source PR: https://github.com/microsoft/tabster/pull/549

Copilot-Session: beb4e74e-7507-444b-b9fd-bfcce9943b93
