+++ src/get/getRootDummyInputs.ts
@@ -5,21 +5,26 @@
 
 import { DummyInputManager, ensureDummyInputObserver } from "../DummyInput.js";
 import { createRootDummyManager } from "../RootDummyManager.js";
+import { installTabKeyHandler } from "../Tab.js";
 import type * as Types from "../Types.js";
 
 /**
- * Opt the Tabster instance into root-level dummy-input behaviour.
- * Registers the root dummy-manager factory plus the dummy-input observer
- * and the `moveOutOfRoot` routing, then adds root-level dummy inputs to
- * existing roots. Per-part dummy factories (Mover, Groupper, Modalizer)
- * are registered by their respective `getX` factories, so consumers only
- * pay for the dummy code of features they actually use.
+ * Opt the Tabster instance into Tab-key control + root-level dummy
+ * inputs.
  *
- * `createTabster` calls this automatically when `controlTab` (default
- * `true`) or `rootDummyInputs` is on, so the default
- * `createTabster(window)` "just works." Apps that opt out
- * (`createTabster(win, { controlTab: false })` without calling this) keep
- * the slim, dummy-free baseline.
+ * Calling this function *is* the opt-in. It registers the root
+ * dummy-manager factory, the dummy-input observer, the `moveOutOfRoot`
+ * routing, the Tab-key keydown handler, and adds root-level dummy
+ * inputs to existing roots — all of which used to be eagerly installed
+ * by `createTabster`.
+ *
+ * Per-feature dummies (Mover/Groupper/Modalizer) don't need this:
+ * `getMover`/`getGroupper`/`getModalizer` ensure their own dummy
+ * infrastructure when called.
+ *
+ * Default `createTabster(win)` is the slim baseline (no Tab control, no
+ * root dummies, no phantom-dummy machinery). Pair it with
+ * `getRootDummyInputs(tabster)` for "Tab just works" behaviour.
  */
 export function getRootDummyInputs(tabster: Types.Tabster): void {
     const tabsterCore = tabster.core;
@@ -55,6 +60,15 @@ export function getRootDummyInputs(tabster: Types.Tabster): void {
                 );
             }
         };
+
+        // Install the Tab-key handler. Lives in src/Tab.ts so its
+        // dependencies (`findNextTabbable`, `DummyInputManager`'s phantom
+        // helpers, the keydown wiring) only enter the bundle here.
+        const stopTabKeyHandler = installTabKeyHandler(
+            tabsterCore,
+            tabsterCore.getWindow
+        );
+        tabsterCore.disposers.add({ dispose: stopTabKeyHandler });
 
         // Apply to existing roots; per-part Mover/Groupper/Modalizer
         // dummies are picked up lazily inside the per-part constructors
